+++ sportorg/models/start/relay.py
@@ -33,7 +33,6 @@ def get_next_relay_number_protocol():
 def get_next_relay_number_setting():
     obj = race()
     settings_num = obj.get_setting("relay_next_number", 1001)
-    # protocol_num = get_next_relay_number_protocol()
     return settings_num