+++ sportorg/modules/sfr/sfrreader.py
@@ -29,7 +29,6 @@ class SFRReaderThread(QThread):
 
     def __init__(self, queue, stop_event, logger, debug=False):
         super().__init__()
-        # self.setName(self.__class__.__name__)
         self.setObjectName(self.__class__.__name__)
         self._queue = queue
         self._stop_event = stop_event
@@ -67,7 +66,6 @@ class ResultThread(QThread):
 
     def __init__(self, queue, stop_event, logger, start_time=None):
         super().__init__()
-        # self.setName(self.__class__.__name__)
         self.setObjectName(self.__class__.__name__)
         self._queue = queue
         self._stop_event = stop_event
@@ -152,7 +150,6 @@ class SFRReaderClient:
                 self._queue, self._stop_event, self._logger, debug=True
             )
             self._reader_thread.start()
-        # elif not self._reader_thread.is_alive():
         elif self._reader_thread.isFinished():
             self._reader_thread = None
             self._start_reader_thread()
@@ -165,14 +162,12 @@ class SFRReaderClient:
             if self._call_back:
                 self._result_thread.data_sender.connect(self._call_back)
             self._result_thread.start()
-        # elif not self._result_thread.is_alive():
         elif self._result_thread.isFinished():
             self._result_thread = None
             self._start_result_thread()
 
     def is_alive(self):
         if self._reader_thread and self._result_thread:
-            # return self._reader_thread.is_alive() and self._result_thread.is_alive()
             return (
                 not self._reader_thread.isFinished()
                 and not self._result_thread.isFinished()