+++ sportorg/modules/sportiduino/sportiduino.py
@@ -159,7 +159,6 @@ class SportiduinoClient:
             if self._call_back:
                 self._result_thread.data_sender.connect(self._call_back)
             self._result_thread.start()
-        # elif not self._result_thread.is_alive():
         elif self._result_thread.isFinished():
             self._result_thread = None
             self._start_result_thread()